more refactoring and organization
