+++ glo-visualizer/src/colormaps.ts
@@ -1,0 +1,574 @@
+export const TURBO = [
+    // 0,
+    // 'rgba(33,102,172,0)',
+    0.0,
+    'rgba(48.44, 18.3, 59.2,0)',
+    0.004,
+    'rgb(49.68, 21.26, 66.68)',
+    0.008,
+    'rgb(50.89, 24.22, 74.01)',
+    0.012,
+    'rgb(52.06, 27.16, 81.2)',
+    0.016,
+    'rgb(53.19, 30.1, 88.25)',
+    0.02,
+    'rgb(54.29, 33.01, 95.15)',
+    0.023,
+    'rgb(55.36, 35.92, 101.91)',
+    0.027,
+    'rgb(56.38, 38.82, 108.52)',
+    0.031,
+    'rgb(57.38, 41.7, 114.99)',
+    0.035,
+    'rgb(58.33, 44.58, 121.32)',
+    0.039,
+    'rgb(59.25, 47.44, 127.51)',
+    0.043,
+    'rgb(60.13, 50.29, 133.55)',
+    0.047,
+    'rgb(60.98, 53.12, 139.45)',
+    0.051,
+    'rgb(61.8, 55.95, 145.2)',
+    0.055,
+    'rgb(62.57, 58.76, 150.81)',
+    0.059,
+    'rgb(63.32, 61.56, 156.28)',
+    0.062,
+    'rgb(64.02, 64.35, 161.6)',
+    0.066,
+    'rgb(64.69, 67.13, 166.79)',
+    0.07,
+    'rgb(65.33, 69.9, 171.82)',
+    0.074,
+    'rgb(65.93, 72.65, 176.71)',
+    0.078,
+    'rgb(66.49, 75.4, 181.46)',
+    0.082,
+    'rgb(67.01, 78.13, 186.07)',
+    0.086,
+    'rgb(67.51, 80.85, 190.53)',
+    0.09,
+    'rgb(67.96, 83.56, 194.85)',
+    0.094,
+    'rgb(68.38, 86.25, 199.03)',
+    0.098,
+    'rgb(68.77, 88.94, 203.06)',
+    0.102,
+    'rgb(69.11, 91.61, 206.95)',
+    0.105,
+    'rgb(69.43, 94.27, 210.69)',
+    0.109,
+    'rgb(69.7, 96.92, 214.29)',
+    0.113,
+    'rgb(69.94, 99.56, 217.75)',
+    0.117,
+    'rgb(70.15, 102.18, 221.06)',
+    0.121,
+    'rgb(70.32, 104.8, 224.24)',
+    0.125,
+    'rgb(70.45, 107.4, 227.26)',
+    0.129,
+    'rgb(70.55, 109.99, 230.15)',
+    0.133,
+    'rgb(70.61, 112.57, 232.89)',
+    0.137,
+    'rgb(70.64, 115.14, 235.48)',
+    0.141,
+    'rgb(70.63, 117.69, 237.94)',
+    0.145,
+    'rgb(70.58, 120.24, 240.25)',
+    0.148,
+    'rgb(70.5, 122.77, 242.41)',
+    0.152,
+    'rgb(70.39, 125.29, 244.44)',
+    0.156,
+    'rgb(70.23, 127.79, 246.31)',
+    0.16,
+    'rgb(70.05, 130.29, 248.05)',
+    0.164,
+    'rgb(69.82, 132.78, 249.64)',
+    0.168,
+    'rgb(69.55, 135.25, 251.08)',
+    0.172,
+    'rgb(69.12, 137.74, 252.27)',
+    0.176,
+    'rgb(68.54, 140.24, 253.22)',
+    0.18,
+    'rgb(67.81, 142.75, 253.94)',
+    0.184,
+    'rgb(66.94, 145.27, 254.42)',
+    0.188,
+    'rgb(65.95, 147.79, 254.68)',
+    0.191,
+    'rgb(64.83, 150.32, 254.73)',
+    0.195,
+    'rgb(63.61, 152.85, 254.58)',
+    0.199,
+    'rgb(62.29, 155.39, 254.23)',
+    0.203,
+    'rgb(60.88, 157.92, 253.69)',
+    0.207,
+    'rgb(59.38, 160.45, 252.97)',
+    0.211,
+    'rgb(57.82, 162.98, 252.07)',
+    0.215,
+    'rgb(56.2, 165.5, 251.01)',
+    0.219,
+    'rgb(54.52, 168.01, 249.8)',
+    0.223,
+    'rgb(52.81, 170.51, 248.43)',
+    0.227,
+    'rgb(51.05, 173.0, 246.92)',
+    0.23,
+    'rgb(49.28, 175.47, 245.28)',
+    0.234,
+    'rgb(47.49, 177.93, 243.52)',
+    0.238,
+    'rgb(45.7, 180.37, 241.64)',
+    0.242,
+    'rgb(43.92, 182.78, 239.65)',
+    0.246,
+    'rgb(42.15, 185.18, 237.56)',
+    0.25,
+    'rgb(40.4, 187.56, 235.38)',
+    0.254,
+    'rgb(38.69, 189.9, 233.11)',
+    0.258,
+    'rgb(37.02, 192.22, 230.76)',
+    0.262,
+    'rgb(35.41, 194.51, 228.35)',
+    0.266,
+    'rgb(33.86, 196.77, 225.88)',
+    0.27,
+    'rgb(32.38, 198.99, 223.35)',
+    0.273,
+    'rgb(30.99, 201.18, 220.78)',
+    0.277,
+    'rgb(29.68, 203.34, 218.18)',
+    0.281,
+    'rgb(28.48, 205.45, 215.54)',
+    0.285,
+    'rgb(27.38, 207.52, 212.88)',
+    0.289,
+    'rgb(26.41, 209.55, 210.21)',
+    0.293,
+    'rgb(25.57, 211.54, 207.54)',
+    0.297,
+    'rgb(24.86, 213.47, 204.87)',
+    0.301,
+    'rgb(24.31, 215.36, 202.21)',
+    0.305,
+    'rgb(23.91, 217.2, 199.57)',
+    0.309,
+    'rgb(23.68, 218.98, 196.96)',
+    0.312,
+    'rgb(23.63, 220.71, 194.39)',
+    0.316,
+    'rgb(23.77, 222.39, 191.85)',
+    0.32,
+    'rgb(24.1, 224.0, 189.38)',
+    0.324,
+    'rgb(24.64, 225.56, 186.96)',
+    0.328,
+    'rgb(25.39, 227.05, 184.6)',
+    0.332,
+    'rgb(26.37, 228.48, 182.32)',
+    0.336,
+    'rgb(27.58, 229.86, 180.03)',
+    0.34,
+    'rgb(29.0, 231.22, 177.61)',
+    0.344,
+    'rgb(30.64, 232.54, 175.08)',
+    0.348,
+    'rgb(32.47, 233.84, 172.45)',
+    0.352,
+    'rgb(34.49, 235.1, 169.72)',
+    0.355,
+    'rgb(36.7, 236.33, 166.89)',
+    0.359,
+    'rgb(39.07, 237.54, 163.99)',
+    0.363,
+    'rgb(41.61, 238.7, 161.0)',
+    0.367,
+    'rgb(44.31, 239.84, 157.94)',
+    0.371,
+    'rgb(47.15, 240.93, 154.82)',
+    0.375,
+    'rgb(50.13, 242.0, 151.64)',
+    0.379,
+    'rgb(53.24, 243.03, 148.41)',
+    0.383,
+    'rgb(56.46, 244.01, 145.13)',
+    0.387,
+    'rgb(59.79, 244.97, 141.82)',
+    0.391,
+    'rgb(63.23, 245.88, 138.47)',
+    0.395,
+    'rgb(66.76, 246.75, 135.1)',
+    0.398,
+    'rgb(70.37, 247.58, 131.72)',
+    0.402,
+    'rgb(74.06, 248.38, 128.32)',
+    0.406,
+    'rgb(77.81, 249.13, 124.92)',
+    0.41,
+    'rgb(81.62, 249.83, 121.52)',
+    0.414,
+    'rgb(85.47, 250.5, 118.13)',
+    0.418,
+    'rgb(89.36, 251.12, 114.76)',
+    0.422,
+    'rgb(93.28, 251.69, 111.4)',
+    0.426,
+    'rgb(97.22, 252.22, 108.08)',
+    0.43,
+    'rgb(101.18, 252.7, 104.8)',
+    0.434,
+    'rgb(105.13, 253.13, 101.56)',
+    0.438,
+    'rgb(109.08, 253.52, 98.37)',
+    0.441,
+    'rgb(113.02, 253.86, 95.23)',
+    0.445,
+    'rgb(116.93, 254.14, 92.16)',
+    0.449,
+    'rgb(120.81, 254.38, 89.16)',
+    0.453,
+    'rgb(124.64, 254.56, 86.23)',
+    0.457,
+    'rgb(128.42, 254.69, 83.39)',
+    0.461,
+    'rgb(132.15, 254.77, 80.64)',
+    0.465,
+    'rgb(135.8, 254.79, 77.98)',
+    0.469,
+    'rgb(139.38, 254.76, 75.43)',
+    0.473,
+    'rgb(142.87, 254.68, 72.99)',
+    0.477,
+    'rgb(146.26, 254.53, 70.67)',
+    0.48,
+    'rgb(149.55, 254.33, 68.46)',
+    0.484,
+    'rgb(152.72, 254.08, 66.4)',
+    0.488,
+    'rgb(155.77, 253.76, 64.46)',
+    0.492,
+    'rgb(158.69, 253.38, 62.68)',
+    0.496,
+    'rgb(161.47, 252.95, 61.04)',
+    0.5,
+    'rgb(164.12, 252.45, 59.56)',
+    0.504,
+    'rgb(166.75, 251.88, 58.23)',
+    0.508,
+    'rgb(169.39, 251.24, 57.04)',
+    0.512,
+    'rgb(172.03, 250.53, 56.0)',
+    0.516,
+    'rgb(174.66, 249.75, 55.09)',
+    0.52,
+    'rgb(177.29, 248.91, 54.3)',
+    0.523,
+    'rgb(179.91, 248.0, 53.63)',
+    0.527,
+    'rgb(182.52, 247.03, 53.08)',
+    0.531,
+    'rgb(185.12, 246.0, 52.63)',
+    0.535,
+    'rgb(187.71, 244.91, 52.29)',
+    0.539,
+    'rgb(190.27, 243.76, 52.04)',
+    0.543,
+    'rgb(192.82, 242.56, 51.87)',
+    0.547,
+    'rgb(195.35, 241.3, 51.79)',
+    0.551,
+    'rgb(197.86, 239.99, 51.79)',
+    0.555,
+    'rgb(200.34, 238.63, 51.86)',
+    0.559,
+    'rgb(202.79, 237.21, 51.98)',
+    0.562,
+    'rgb(205.21, 235.75, 52.17)',
+    0.566,
+    'rgb(207.6, 234.25, 52.41)',
+    0.57,
+    'rgb(209.95, 232.7, 52.69)',
+    0.574,
+    'rgb(212.26, 231.1, 53.01)',
+    0.578,
+    'rgb(214.54, 229.46, 53.36)',
+    0.582,
+    'rgb(216.78, 227.79, 53.74)',
+    0.586,
+    'rgb(218.96, 226.07, 54.14)',
+    0.59,
+    'rgb(221.11, 224.32, 54.55)',
+    0.594,
+    'rgb(223.2, 222.53, 54.97)',
+    0.598,
+    'rgb(225.24, 220.71, 55.38)',
+    0.602,
+    'rgb(227.24, 218.86, 55.79)',
+    0.605,
+    'rgb(229.17, 216.97, 56.2)',
+    0.609,
+    'rgb(231.04, 215.06, 56.58)',
+    0.613,
+    'rgb(232.86, 213.12, 56.94)',
+    0.617,
+    'rgb(234.61, 211.16, 57.26)',
+    0.621,
+    'rgb(236.3, 209.16, 57.55)',
+    0.625,
+    'rgb(237.92, 207.15, 57.8)',
+    0.629,
+    'rgb(239.47, 205.12, 58.0)',
+    0.633,
+    'rgb(240.95, 203.07, 58.14)',
+    0.637,
+    'rgb(242.35, 201.0, 58.22)',
+    0.641,
+    'rgb(243.68, 198.91, 58.23)',
+    0.645,
+    'rgb(244.92, 196.81, 58.17)',
+    0.648,
+    'rgb(246.09, 194.7, 58.02)',
+    0.652,
+    'rgb(247.17, 192.57, 57.79)',
+    0.656,
+    'rgb(248.17, 190.44, 57.47)',
+    0.66,
+    'rgb(249.08, 188.3, 57.04)',
+    0.664,
+    'rgb(249.9, 186.15, 56.51)',
+    0.668,
+    'rgb(250.64, 183.96, 55.89)',
+    0.672,
+    'rgb(251.3, 181.69, 55.21)',
+    0.676,
+    'rgb(251.89, 179.34, 54.46)',
+    0.68,
+    'rgb(252.41, 176.92, 53.66)',
+    0.684,
+    'rgb(252.87, 174.44, 52.8)',
+    0.688,
+    'rgb(253.25, 171.89, 51.89)',
+    0.691,
+    'rgb(253.57, 169.28, 50.93)',
+    0.695,
+    'rgb(253.81, 166.62, 49.92)',
+    0.699,
+    'rgb(254.0, 163.91, 48.87)',
+    0.703,
+    'rgb(254.12, 161.14, 47.78)',
+    0.707,
+    'rgb(254.17, 158.34, 46.66)',
+    0.711,
+    'rgb(254.16, 155.49, 45.5)',
+    0.715,
+    'rgb(254.09, 152.61, 44.31)',
+    0.719,
+    'rgb(253.96, 149.69, 43.09)',
+    0.723,
+    'rgb(253.77, 146.75, 41.85)',
+    0.727,
+    'rgb(253.52, 143.78, 40.59)',
+    0.73,
+    'rgb(253.21, 140.8, 39.31)',
+    0.734,
+    'rgb(252.84, 137.79, 38.02)',
+    0.738,
+    'rgb(252.42, 134.78, 36.71)',
+    0.742,
+    'rgb(251.94, 131.75, 35.4)',
+    0.746,
+    'rgb(251.4, 128.72, 34.09)',
+    0.75,
+    'rgb(250.82, 125.69, 32.76)',
+    0.754,
+    'rgb(250.18, 122.67, 31.45)',
+    0.758,
+    'rgb(249.48, 119.65, 30.13)',
+    0.762,
+    'rgb(248.74, 116.64, 28.83)',
+    0.766,
+    'rgb(247.95, 113.64, 27.53)',
+    0.77,
+    'rgb(247.11, 110.67, 26.25)',
+    0.773,
+    'rgb(246.22, 107.71, 24.98)',
+    0.777,
+    'rgb(245.28, 104.79, 23.74)',
+    0.781,
+    'rgb(244.29, 101.89, 22.52)',
+    0.785,
+    'rgb(243.26, 99.03, 21.32)',
+    0.789,
+    'rgb(242.19, 96.21, 20.16)',
+    0.793,
+    'rgb(241.07, 93.43, 19.03)',
+    0.797,
+    'rgb(239.91, 90.69, 17.93)',
+    0.801,
+    'rgb(238.71, 88.01, 16.87)',
+    0.805,
+    'rgb(237.47, 85.38, 15.86)',
+    0.809,
+    'rgb(236.19, 82.81, 14.88)',
+    0.812,
+    'rgb(234.87, 80.3, 13.96)',
+    0.816,
+    'rgb(233.51, 77.85, 13.09)',
+    0.82,
+    'rgb(232.11, 75.48, 12.28)',
+    0.824,
+    'rgb(230.68, 73.17, 11.52)',
+    0.828,
+    'rgb(229.21, 70.95, 10.82)',
+    0.832,
+    'rgb(227.71, 68.8, 10.18)',
+    0.836,
+    'rgb(226.16, 66.69, 9.57)',
+    0.84,
+    'rgb(224.57, 64.6, 8.98)',
+    0.844,
+    'rgb(222.93, 62.54, 8.41)',
+    0.848,
+    'rgb(221.24, 60.51, 7.86)',
+    0.852,
+    'rgb(219.5, 58.51, 7.33)',
+    0.855,
+    'rgb(217.72, 56.53, 6.83)',
+    0.859,
+    'rgb(215.89, 54.59, 6.34)',
+    0.863,
+    'rgb(214.01, 52.67, 5.88)',
+    0.867,
+    'rgb(212.09, 50.78, 5.43)',
+    0.871,
+    'rgb(210.12, 48.91, 5.01)',
+    0.875,
+    'rgb(208.1, 47.08, 4.61)',
+    0.879,
+    'rgb(206.04, 45.27, 4.23)',
+    0.883,
+    'rgb(203.93, 43.49, 3.88)',
+    0.887,
+    'rgb(201.77, 41.74, 3.54)',
+    0.891,
+    'rgb(199.56, 40.02, 3.22)',
+    0.895,
+    'rgb(197.31, 38.32, 2.93)',
+    0.898,
+    'rgb(195.01, 36.65, 2.65)',
+    0.902,
+    'rgb(192.67, 35.01, 2.4)',
+    0.906,
+    'rgb(190.27, 33.4, 2.17)',
+    0.91,
+    'rgb(187.84, 31.82, 1.96)',
+    0.914,
+    'rgb(185.35, 30.26, 1.77)',
+    0.918,
+    'rgb(182.81, 28.73, 1.6)',
+    0.922,
+    'rgb(180.23, 27.23, 1.46)',
+    0.926,
+    'rgb(177.61, 25.76, 1.33)',
+    0.93,
+    'rgb(174.94, 24.32, 1.23)',
+    0.934,
+    'rgb(172.21, 22.9, 1.14)',
+    0.938,
+    'rgb(169.44, 21.51, 1.08)',
+    0.941,
+    'rgb(166.63, 20.15, 1.04)',
+    0.945,
+    'rgb(163.77, 18.82, 1.02)',
+    0.949,
+    'rgb(160.86, 17.51, 1.02)',
+    0.953,
+    'rgb(157.9, 16.24, 1.05)',
+    0.957,
+    'rgb(154.9, 14.99, 1.09)',
+    0.961,
+    'rgb(151.85, 13.77, 1.16)',
+    0.965,
+    'rgb(148.76, 12.57, 1.24)',
+    0.969,
+    'rgb(145.61, 11.41, 1.35)',
+    0.973,
+    'rgb(142.42, 10.27, 1.48)',
+    0.977,
+    'rgb(139.19, 9.16, 1.63)',
+    0.98,
+    'rgb(135.9, 8.08, 1.8)',
+    0.984,
+    'rgb(132.57, 7.03, 1.99)',
+    0.988,
+    'rgb(129.19, 6.0, 2.2)',
+    0.992,
+    'rgb(125.77, 5.01, 2.44)',
+    0.996,
+    'rgb(122.3, 4.04, 2.69)'
+]
+
+export const VIRIDIS = [0.0,
+    'rgba(68, 1, 84,0)',
+    0.01,
+    'rgb(68, 1, 84)',
+    0.13,
+    'rgb(71, 44, 122)',
+    0.25,
+    'rgb(59, 81, 139)',
+    0.38,
+    'rgb(44, 113, 142)',
+    0.5,
+    'rgb(33, 144, 141)',
+    0.63,
+    'rgb(39, 173, 129)',
+    0.75,
+    'rgb(92, 200, 99)',
+    0.88,
+    'rgb(170, 220, 50)',
+    1.0,
+    'rgb(253, 231, 37)'
+];
+
+export const CUBE_HELIX = [0.0,
+    'rgba(0, 0, 0,0)',
+    0.01,
+    'rgb(0, 0, 0)',
+    0.0175,
+    'rgb(22, 5, 59)',
+    0.0325,
+    'rgb(60, 4, 105)',
+    0.05,
+    'rgb(109, 1, 135)',
+    0.0675,
+    'rgb(161, 0, 147)',
+    0.0825,
+    'rgb(210, 2, 142)',
+    0.1,
+    'rgb(251, 11, 123)',
+    0.1175,
+    'rgb(255, 29, 97)',
+    0.1325,
+    'rgb(255, 54, 69)',
+    0.15,
+    'rgb(255, 85, 46)',
+    0.1675,
+    'rgb(255, 120, 34)',
+    0.1825,
+    'rgb(255, 157, 37)',
+    0.2,
+    'rgb(241, 191, 57)',
+    0.2175,
+    'rgb(224, 220, 93)',
+    0.2325,
+    'rgb(218, 241, 142)',
+    0.25,
+    'rgb(227, 253, 198)'
+];
